chore: extract and rename merge => reconstruct
